Backup system has been refactored
1. Now there's an option to save the state before the generation
update, after it or at both moments.

2. To guarantee that the behavior of continue.py is identical to
the behavior of main.py after the restoration, all settings
governing output have been moved to the Evolver's params. Calls to
the corresponding functions are now mandatory, but the functions
are rewritten to check whether the parameters are set and run only
if they are.

3. Small bug is fixed: previously continue.py started its own log
for best individuals instead of adding to the old one.

	modified:   continue.py
	modified:   evolvers/baseEvolver.py
	modified:   main.py