Optimization of the behavior of Pareto finder for brute forcers
I noticed that it spends N^2 time at each generation, where N is
a number of all individuals known to be at the Pareto front at the
time, INCLUDING those with coinciding objective pairs. Rewrote it
to require only M^2, where M is the number of all known unique
objective pairs. The cost is a bunch of operations linear in N,
but I think it will be well worth it.

	modified:   evolvers/bruteForcePareto.py