Removed the member variable objective function
Before this commit, evolvers/bruteForcePareto.py used a member
variable to store a callable used to produce the second
optimization objective. Objects which store functions in this
manner are not suitable for pickling via regular means, which
led me to the replace all function calls with calls to its
explicit lambda.

	modified:   evolvers/bruteForcePareto.py